server/requirements.txt
Flask==1.0.3
torch==1.3.1
torchvision==0.4.2
Pillow==5.2.0